[PA][BRP] Obliterate support for extras offset
It was used only for BRP "before allocation" mode, which is now gone
(crrev.com/c/5117767). And we aren't planning to go back that route.

Previously we used it also for a leading cookie in debug builds. But
that just complicated code, and is long gone too.

Having any leading extras requires setting up a separate partition for
aligned allocations, because they conflict with each other. But that
support is now gone as well (crrev.com/c/5119422).

Bug: 1511221
Change-Id: I43839d2ed41a2cd5db29f048373bb25d274fb175
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5119424
Reviewed-by: Mikihito Matsuura <mikt@google.com>
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1237935}
NOKEYCHECK=True
GitOrigin-RevId: bcfc1b43b2ff0d3459063b390cdf3fba1f0f3323