[BRP] Implement "same slot" mode
It's currently behind the bulidflag (!put_ref_count_in_previous_slot),
but the intention is to make it selectable at run-time, in order to
finch it.

Ref-count in the previous slot suffers from a race, when that slot
is freed and its MTE tag changes. Putting the ref-count in the same
slot avoids that entirely.

Ref-count in the same slot has another advantage of not sharing the
slot with the freelist pointer, so we no longer have to worry about
sizes. This CL keeps this restriction, however, as it'll make it
easier to convert it into a run-time flag.

The "same slot" mode resembles the "after allocation" mode we had prior
to crrev.com/c/2783963. The difference is that ref-count would be
placed immediately after the allocation (when raw size is known, i.e.
for direct map and single-slot spans), which added cycle to ref-count
address calculation, and created the need for ref-count relocation
on realloc.

It felt like there was an opportunity to get rid of the ref-count
table (tucked in the super page metadata area), and thus a branch
in PartitionRefCountPointer(). But there were multiple issues (see
the lengthy comment in that function), that the easiest way to work
around would be by keeping the ref-count table.

Bug: 1511221, 1445816
Change-Id: If128d46b38d775d5167e629f88a594e2cd23b304
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5125433
Reviewed-by: Sergei Glazunov <glazunov@google.com>
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1241806}
NOKEYCHECK=True
GitOrigin-RevId: e7bc33a3d203969d55beb6b68785d286e03666b7