DPD: Require feature for flag enablement
There doesn't appear to be any mechanism that requires compiled DPD
support to default-enable the feature flag. This CL is intended to be
trivial (because all of our bots are surely configured correctly
already) and asserts that it's meaningless to default-enable the feature
when the backing code isn't compiled.

Change-Id: I311c4598b050ae647520a34139bc00d329b56432
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5227683
Auto-Submit: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Commit-Queue: Kalvin Lee <kdlee@chromium.org>
Reviewed-by: Arthur Sonzogni <arthursonzogni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1251324}
NOKEYCHECK=True
GitOrigin-RevId: ec3226f3a7ee608ac147725e54c775c5b6748cef