PA: Fix nullptr bug in FreelistDispatcher
In the following revert,
https://chromium-review.googlesource.com/c/chromium/src/+/5408134
enabling the pool offset freelist feature leads to
undefined behavior when a nullptr is passed in.

In order to fix undefined behavior with the feature is enabled,
any instances where we use the `next` pointer are cast
converted into the pointerof the specific freelist type. Thus,
`nullptr` can be passed in as a parameter and still be used
to access the freelist type.

Bug: 40274683
Change-Id: If7d06becc5e8425dbc14cd1f38939d3295d72964
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5411834
Auto-Submit: Ada Romanowski <adarom@google.com>
Reviewed-by: Bartek Nowierski <bartekn@chromium.org>
Reviewed-by: Mikihito Matsuura <mikt@google.com>
Commit-Queue: Bartek Nowierski <bartekn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1282386}
NOKEYCHECK=True
GitOrigin-RevId: cdde234798dce22b3432b1ea53f86d06f6063f2a