[PA] Split PartitionSuperPageExtentEntry into read-only one and writable one.
This is a part of ShadowMetadata CL (2/5). The entire CL is
crrev.com/c/5570783.

ReadOnlyPartitionSuperPageExtentEntry is used for
PartitionSuperPageExtentEntry inside the giga cage, and
WritablePartitionSuperPageExtentEntry is used for
PartitionSuperPageExtentEntry inside the shadow pool.

The design doc of shadow metadata is
https://docs.google.com/document/d/1LvMI_c1VuT4sWUqgKOdJlrZv0UWP5e3PLJ_oOlJ4l2k

Bug: 40238514
No-Try: true
Change-Id: Ib06c7743701b98e168c49781d2d933789a144590
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5684975
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Commit-Queue: Takashi Sakamoto <tasak@google.com>
Cr-Commit-Position: refs/heads/main@{#1337653}
NOKEYCHECK=True
GitOrigin-RevId: 7cddc390c898096526a3911df8b97ac6bfd14bd8