[PartitionAlloc] Name anonymous VMAs on Linux and Chrome OS
The right constants are not necessarily defined at build time, but may
be understood by the kernel. Follow a similar approach to how the
sandbox code defines them.

Note that this does not mean that names are available in practice, since
prctl() may be blocked by some kernel security module, or
CONFIG_ANON_VMA_NAME may not be set (as is currently the case for Debian
kernels).

Bug: 385400561
Change-Id: I674a977abf5c84a22fe6d9afc1f7128c6c75786d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6148169
Commit-Queue: Benoit Lize <lizeb@chromium.org>
Reviewed-by: Keishi Hattori <keishi@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1402852}
NOKEYCHECK=True
GitOrigin-RevId: 3435dc80817bb785bd750d91d423f54d4a42cd82