[PA] Fix unused `use_libcxx_modules`.
The GN args `is_debug` and `use_libcxx_modules` are parts of Chromium.
The second was recently added by:
https://chromium-review.googlesource.com/c/chromium/src/+/6768059

This prevents configuring partition_alloc when building it as a
standalone project, or when building it as part of Skia standalone.

This patch fixes the error:
    ```
    ERROR at //src/partition_alloc/BUILD.gn:592:30: Assignment had no effect.
        use_libcxx_modules = false
    ```

Bug: 428778692
Bug: 351867706
Change-Id: I0f55511f5f010f3c1e50fff25ce8b5f2560d3519
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6791465
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Arthur Sonzogni <arthursonzogni@chromium.org>
Reviewed-by: Takuto Ikuta <tikuta@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1492610}
NOKEYCHECK=True
GitOrigin-RevId: b9e35281a0868e8293e99079fd49e997c6fda8f9